+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * `NegatedIf` properly handles negated `unless` condition. ([@bbatsov][])
 * `NegatedWhile` properly handles negated `until` condition. ([@bbatsov][])
 * [#925](https://github.com/bbatsov/rubocop/issues/925): Do not disable the `Syntax` cop in output from `--auto-gen-config`. ([@jonas054][])
+* [#943](https://github.com/bbatsov/rubocop/issues/943): Fix auto-correction interference problem between `SpaceAfterComma` and other cops. ([@jonas054][])
 
 ## 0.20.0 (02/04/2014)